+++ src/nws_exporter/client.rs
@@ -61,7 +61,7 @@ pub struct NwsClient {
 
 impl NwsClient {
     const USER_AGENT: &'static str =
-        "nws_exporter/0.7.0 (https://github.com/jgrig472/nws_exporter-with-noaa-water-gauge-support)";
+        "nws_exporter/0.8.0 (https://github.com/jgrig472/nws_exporter-with-noaa-water-gauge-support)";
     const JSON_RESPONSE: &'static str = "application/geo+json";
 
     /// Create a new `NwsClient` from the provided reqwest client and based URL for the